Sliding Window Pattern #1
Given an array, find the average of all contiguous subarrays of size ‘K’ in it.

Given an array of positive numbers and a positive number ‘k’, find the maximum sum of any contiguous subarray of size ‘k’.

Given an array of positive numbers and a positive number ‘S’, find the length of the smallest contiguous subarray whose sum is greater than or equal to ‘S’. Return 0, if no such subarray exists. This uses sliding window technique without a fixed size of the window.

Given a string, find the length of the longest substring in it with no more than K distinct characters.

Given a string, find the length of the longest substring which has no repeating characters.

Given a string with lowercase letters only, if you are allowed to replace no more than ‘k’ letters with any letter, find the length of the longest substring having the same letters after replacement.

Given a string and a pattern, find out if the string contains any permutation of the pattern. This can also be used to find all the permutations in the string.